Add tests for Stripe checkout w/ hiring agreements
